MCP: move cancellation test tools to their own app
